[refactor]: clean up warnings and qos type



# message auto-generated for no-merge-commit merge:
!609 merge refactor/cleanup-hixl-warnings into master

[refactor]: clean up warnings and qos type

Created-by: zhaozhihui
Commit-by: zhaozhihui
Merged-by: cann-robot
Description: **提交 PR 前，请先阅读 [HIXL贡献指南](https://gitcode.com/cann/hixl/blob/master/CONTRIBUTING.md) ，查看 PR 提交方式并遵守相关提交规范。**

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [x] 代码重构
- [x] 文档更新
- [ ] 其他，请描述：

## 描述
<!--简要描述本次改动的背景，包括改动的原因、解决的问题等。-->

- 清理 HIXL 相关 warning，补齐接口转换和日志格式。
- 将 qos 在 HIXL 内部配置、控制消息、channel 描述和 handler 参数链路调整为 uint8_t。
- hcomm 的 qos 定义保持 uint32_t 不变，传入 hcomm 时继续显式转换。
- 更新 HIXL 接口和 HIXL CS 接口文档中的 qos 配置说明与示例，CS qos 描述与 HIXL 接口文档保持一致。

## 测试项
<!--描述进行了哪些测试来验证本次改动，或新增了哪些测试用例。-->

已有用例

## 测试结果
<!--描述上述测试项的测试结果，可通过表格、图片等形式展示。-->

PASS

## Checklist
<!--  [x] 表示选中 -->
- [x] 该Pull Request的代码风格和项目的代码风格一致
- [x] 提交的代码已经过充分的验证
- [x] 涉及的相关文档修改已更新
- [x] 标题中正确使用了类型标签（例如：feat/bugfix/refactor/docs/test等）

## 其它（可选）
<!--可以在这里描述任何和该Pull Request相关的说明。-->

- 本次未跑全量用例；测试结果仅呈现与 qos 类型调整和文档修改相关的检查。


See merge request: cann/hixl!609

+++ src/hixl/engine/hixl_client.h
@@ -32,7 +32,7 @@ struct ClientConfig {
   uint8_t rdma_sl;
   uint32_t timeout_ms;
   std::optional<uint32_t> local_listen_port;
-  std::optional<int32_t> qos;
+  std::optional<uint8_t> qos;
 };
 
 class HixlClient {
@@ -144,7 +144,7 @@ class HixlClient {
                              // inflight 配对
   std::mutex req_map_mutex_;
   std::map<TransferReq, TransferInfo> req_map_;
-  std::optional<int32_t> qos_;
+  std::optional<uint8_t> qos_;
 };
 
 }  // namespace hixl
